requirements.txt
discord.py[voice]==0.16.12
mysql-connector-python==8.0.17
py-rolldice==0.3.9
pytz==2018.5